Adds buttons to a transaction history row for easier use.
Adds an 'Abandon Transaction' button to the row header to more quicker abandon a transaction.
While the button was available in the expanded history, this adds the button to the parent row for convenience.
Additionally, allows for abandoning an internal transfer (conversion) transction, rather than just send transactions.

Adds a button to allow for copying of the transaction address (the address to/from which the transaction targets).
The ability to copy the address already exists via double-click to highlight the visible part of the address, ctrl/cmd-C.
This however, is hidden from GUI users, so a button has been added for this purpose.
While right-click style context menus would be useful, those suffer from their own issues:
  - has similar discoverability issues (although maybe to a slightly lesser extent) as the current double-click ctrl-C style copy;
  - is more complicated to include as a feature currently due to electron and safely integrating node and related 'backend' API directly into the window.
The context menus can be investigated as part of a future discussion.