[Market] Removes the market browser and promotion of markets.
Current legal proceedings suggest that the Particl team may be held accountable for illegal/iilicit materials sold via markets made visible/known via the application.
The relevant legal proceedings are still ongoing, and the extent of the impact, if any, on the Particl related software is still to be determined.
However, a pre-emptive approach appears to be better longer term here than a reactive one, particularly giving the nature of various items already visible in various promoted markets already.

On the surface, this effectively boils down to moderation of content, and prevention of adverse listings from reaching a user's machine for openly public/promoted markets. Thus, there is active exploration into various alternative means of market moderation for publicly visible markets via a decentralized consensus mechanism (which in no way should enforce specific beliefs or impose any particular constraints, and the goals and aims of Particl reamin as is). However, while such solutions are being investigated and tested, potential liability from a legal standpoint still needs to be considered.

To this extent, this change removes the market browser and ability to promote a market via the Particl Desktop user interface.
This is intended to be a temporary measure, at least until such a solution as mentioned above is implemented and available.